requirements.txt
audioread==3.0.0
cmpfilter
dfcon
easy-video
joblib==1.3.2
librosa==0.10.1
matplotlib==3.7.3
mediapipe
numpy==1.24.4
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74
soundfile==0.12.1
toolpack 
pyworld
librosa 
wandb
torchaudio
pytorch-lightning
moviepy